[doc] Add C Style Guide Rules for Symbol Naming
Unlike C++, C does not have user-defined namespaces. This means
programmers need to be a lot more careful when naming types, functions,
and symbols.

The following two rules have come up in the software team when
discussing the DIFs, and I feel they are important enough to go into the
OpenTitan-wide C style guide.

1. All symbols in a particular header must share the same unique prefix.
2. The names of enumeration constants must be prefixed with the name of
   their respective enumeration type.

This commit adds both rules with some explanation as to why they are
needed, and some illustrative examples.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>