[doc] Document Top Earlgrey Parameterisations
At the moment, we only have one top level (top_earlgrey) and therefore
only one parameterisation of various peripherals (alert handler, pinmux,
PLIC).

The docs, until today, have been generated from the "default"
parameterisations, which have not matched the top earlgrey
parameterisations, which has made writing software against the hardware
difficult and error-prone.

This commit changes the docs to use the earlgrey parameterisations, and
documents the parameterisations in the parameter tables. Should we gain
other top levels, we should revisit how this is generated, but we knew
that already, and this makes the software developers lives much easier.

Closes #1068, #1496

Signed-off-by: Sam Elliott <selliott@lowrisc.org>