[CI] Ensure apt package index is up-to-date
Installing packages with `apt-get` requires the package index to
up-to-date. If not, the package installation might fail due to packages
being selected for installation which don't exist any more.

In hosted CI runners, this step is automatically taken care of, and
adding an explicit `apt-get update` is a no-op (except for taking a
small amount of time). In other environments, it's essential for
correctness.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>