[sw] Fix printf not working in RISC-V
This is an unfortunate side-effect of the difference between va_list on
amd64 and rv32imc. On the former, it is a double pointer, so advancing
it in process_specifier() actually advances it overall. On rv32imc,
however, it's just a void *, so a pointer to it needs to be passed.

This commit changes print.c to correctly advance the va_list in rv32imc,
by passing a pointer. This incurs a copy on amd64, but is free on
rv32imc.

Signed-off-by: Miguel Young de la Sota <mcyoung@google.com>