[sw, spiflash] Dramatically increase transfer rate
The spiflash tool has relied on a fixed delay for spacing the writes
between frames. This hasn't been a problem for small binary deployments
to the FPGA, but Tock release binaries take ~3m23s to flash, and debug
binaries take even longer.

Add flow control to spiflash by waiting for until the hash of the
current frame can be read back before writing the next frame. This
successfully allows Tock to be flashed to the FPGA in 6.7s, a 96.7%
decrease over the previous implementation.

Signed-off-by: Jon Flatley <jflat@google.com>