[reggen] Fix Interrupt `genresmask`
Problem:

    Multireg doesn't show correct mask value

The auto generated multi registers doesn't consider the multi-width
field in the interrupt related registers. Now the code is changed to
consider the `width` field to correctly calculate the `genresmask`.

Also in multireg validation code, when it creates a `genreg`, it didn't
store the `genresval` and `genresmask`. So the register table in the
specification only showed 0 reset value and 1 mask value.

This fixes #41

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>