[dv, sw] UART logs to UVM prints
This PR enables logs from SW test sent over UART to be captured as UVM
print messages. The printed message looks for severity identifier in the
message string (INFO, WARN, ERROR and FATAL) and prints the final
message with the corresponding uvm_* facilities.

The plan to enable the much faster print-bypassing-uart is still on -
that will be enabled shortly. The goal is to have both methods supported
so that FPGA tests can be run as-is in DV without any modifications.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>