[topgen] Array of inter-module signal
This commit is to support an array of inter-module signals.

To achieve the array feature in inter-module, regardless of the act
(requester, responder/receiver), the key shall be an array and the
entries should be `width` as 1. So, inter-module signal only supports
1:1, 1:N, N:1.

Also, the top-level signal name follows the key instance name not
'req->rsp' as before.

```hjson
inter_module: {
  'module_a.sig_a': ['module_b.sig_a', 'module_c.sig_a']
}
```

In the above example, `sig_a` in `module_a` can be requester or
responder/receiver.  The `width` of `module_a.sig_a` shall be 2.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>