[rv_plic] Change PRIO number starting from 1
As discussed in #1823, this PR is to match the PRIO* registers name to
the ID read in CC register.

Previously, ID read from CC register is not matched with the offset of
PRIO0, IS, IE, IP. For example, if the software read 0x3 value from CC
register, its corresponding bit in IS is IS[2].

PLIC now generates ID w.r.t the interrupt source bit position. RISC-V
PLIC spec requires to return ID value 0 when no interrupt. So, topgen
ties bit0 of interrupt source signal to 0. By doing in this way, reggen
doesn't have to support arithmetic operation to the parameter.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>