[CI] Use packaged Verilator
The Verilator package is provided by an experimental repository from
Open Build Service (OBS), which is maintained by me. The repository
differs from "traditional" distribution repositories, as it provides
multiple versioned packages (e.g. verilator-4.032), as opposed to only a
single version. This allows CI to choose and pin exactly one version,
without having to follow the latest releases.

All of this infrastructure is currently experimental and not supported
anywhere except for the use in CI; however, it has been used
successfully in Ibex for multiple months and has proven to work well.

Signed-off-by: Philipp Wagner <phw@lowrisc.org>