[docker] Added clang-format-3.8 to docker
The version of clang-format used in CI is 3.8 which is much older than
most of our machines that run newer linux distros. With this update, one
can build our docker image with clang-format-3.8 and stay consistent
with the CI. IMO this is better than installing multiple packages in our
native machines.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>