[dvsim] Make it simpler to derive from FlowCfg
Before this patch, a subclass had to call

   self.parse_flow_cfg(...)
   self._post_parse_flow_cfg()

Ideally, we wouldn't need either, but subclasses generally need to do
some set-up of their fields before parsing the configuration
file (because of the way that we conflate member variables and hjson
attributes).

This patch simplifies things slightly, so that they can now just call

   self._parse_flow_cfg(...)

(adding an underscore to the name to follow the normal Python naming
convention for protected-ish methods).

To do this, the _parse_flow_cfg() method now calls _parse_cfg (which
recurses if it sees import_cfgs). That means that the body of the old
_post_parse_flow_cfg can be slotted in at the end of _parse_flow_cfg.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>