[topgen] Generate Linker Script Memory Definitions
As pointed out in the reviews of #2031, we can generate the MEMORY
sections of the linker scripts from the information known by topgen.

This commit updates the linker scripts to use this autogenerated
information. Unfortunately there are some slight name changes.

We now only generate a single memory definition (wheras before the flash
linker script did not even mention the ROM, as it is not usable by the
flash image in any reasonable way), which may be a slight drawback.

Signed-off-by: Sam Elliott <selliott@lowrisc.org>